pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-evals"
version = "0.1.0"
description = "Test whether re-priming LLMs with books improves task performance"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "openai>=1.0",
    "pydantic>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "scipy>=1.10",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "datasets>=2.14",
]

[project.scripts]
ai-evals = "ai_evals.cli:cli"

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1"]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_evals"]

[tool.ruff]
line-length = 100
